+++ resources/siege/strats.json
